splitting code tests into table-specific files
