+++ Lib/fontbakery/specifications/glyf_test.py
@@ -1,0 +1,19 @@
+# -*- coding: utf-8 -*-
+from __future__ import absolute_import, print_function, unicode_literals, division
+
+from fontbakery.checkrunner import (
+              DEBUG
+            , INFO
+            , WARN
+            , ERROR
+            , SKIP
+            , PASS
+            , FAIL
+            )
+
+check_statuses = (ERROR, FAIL, SKIP, PASS, WARN, INFO, DEBUG)
+
+# from fontTools.ttLib import TTFont
+
+# TODO: test/069
+
