+++ Lib/fontbakery/specifications/googlefonts_test.py
@@ -1,6 +1,5 @@
 # -*- coding: utf-8 -*-
 from __future__ import absolute_import, print_function, unicode_literals, division
-from builtins import range
 
 import pytest
 
@@ -186,26 +185,6 @@ def test_check_001():
     assert status == FAIL
 
 
-def test_check_002():
-  """ Fonts are all in the same directory. """
-  from fontbakery.specifications.general import com_google_fonts_check_002 as check
-  same_dir = [
-    "data/test/cabin/Cabin-Thin.ttf",
-    "data/test/cabin/Cabin-ExtraLight.ttf"
-  ]
-  multiple_dirs = [
-    "data/test/mada/Mada-Regular.ttf",
-    "data/test/cabin/Cabin-ExtraLight.ttf"
-  ]
-  print('Test PASS with same dir: {}'.format(same_dir))
-  status, message = list(check(same_dir))[-1]
-  assert status == PASS
-
-  print('Test FAIL with multiple dirs: {}'.format(multiple_dirs))
-  status, message = list(check(multiple_dirs))[-1]
-  assert status == FAIL
-
-
 def test_check_003():
   """ Does DESCRIPTION file contain broken links ? """
   from unidecode import unidecode
@@ -306,60 +285,6 @@ def test_check_007():
   assert status == FAIL
 
 
-def test_check_008(mada_ttFonts):
-  """ Fonts have consistent underline thickness ? """
-  from fontbakery.specifications.post import com_google_fonts_check_008 as check
-
-  print('Test PASS with good family.')
-  status, message = list(check(mada_ttFonts))[-1]
-  assert status == PASS
-
-  # introduce a wronge value in one of the font files:
-  value = mada_ttFonts[0]['post'].underlineThickness
-  incorrect_value = value + 1
-  mada_ttFonts[0]['post'].underlineThickness = incorrect_value
-
-  print('Test FAIL with inconsistent family.')
-  status, message = list(check(mada_ttFonts))[-1]
-  assert status == FAIL
-
-
-def test_check_009(mada_ttFonts):
-  """ Fonts have consistent PANOSE proportion ? """
-  from fontbakery.specifications.os2 import com_google_fonts_check_009 as check
-
-  print('Test PASS with good family.')
-  status, message = list(check(mada_ttFonts))[-1]
-  assert status == PASS
-
-  # introduce a wrong value in one of the font files:
-  value = mada_ttFonts[0]['OS/2'].panose.bProportion
-  incorrect_value = value + 1
-  mada_ttFonts[0]['OS/2'].panose.bProportion = incorrect_value
-
-  print('Test FAIL with inconsistent family.')
-  status, message = list(check(mada_ttFonts))[-1]
-  assert status == FAIL
-
-
-def test_check_010(mada_ttFonts):
-  """ Fonts have consistent PANOSE family type ? """
-  from fontbakery.specifications.os2 import com_google_fonts_check_010 as check
-
-  print('Test PASS with good family.')
-  status, message = list(check(mada_ttFonts))[-1]
-  assert status == PASS
-
-  # introduce a wrong value in one of the font files:
-  value = mada_ttFonts[0]['OS/2'].panose.bFamilyType
-  incorrect_value = value + 1
-  mada_ttFonts[0]['OS/2'].panose.bFamilyType = incorrect_value
-
-  print('Test FAIL with inconsistent family.')
-  status, message = list(check(mada_ttFonts))[-1]
-  assert status == FAIL
-
-
 def test_check_011(mada_ttFonts, cabin_ttFonts):
   """ Fonts have equal numbers of glyphs ? """
   from fontbakery.specifications.googlefonts import com_google_fonts_check_011 as check
@@ -393,66 +318,6 @@ def test_check_012(mada_ttFonts, cabin_ttFonts):
   assert status == FAIL
 
 
-def test_check_013(mada_ttFonts):
-  """ Fonts have equal unicode encodings ? """
-  from fontbakery.specifications.cmap import com_google_fonts_check_013 as check
-  from fontbakery.constants import (PLAT_ENC_ID__SYMBOL,
-                                    PLAT_ENC_ID__UCS2)
-  print('Test PASS with good family.')
-  # our reference Mada family is know to be good here.
-  status, message = list(check(mada_ttFonts))[-1]
-  assert status == PASS
-
-  bad_ttFonts = mada_ttFonts
-  # introduce mismatching encodings into the first 2 font files:
-  for i, encoding in enumerate([PLAT_ENC_ID__SYMBOL,
-                                PLAT_ENC_ID__UCS2]):
-    for table in bad_ttFonts[i]['cmap'].tables:
-      if table.format == 4:
-        table.platEncID = encoding
-
-  print('Test FAIL with fonts that diverge on unicode encoding.')
-  status, message = list(check(bad_ttFonts))[-1]
-  assert status == FAIL
-
-
-def test_check_014(mada_ttFonts):
-  """ Make sure all font files have the same version value. """
-  from fontbakery.specifications.head import com_google_fonts_check_014 as check
-
-  print('Test PASS with good family.')
-  # our reference Mada family is know to be good here.
-  status, message = list(check(mada_ttFonts))[-1]
-  assert status == PASS
-
-  bad_ttFonts = mada_ttFonts
-  # introduce a mismatching version value into the second font file:
-  version = bad_ttFonts[0]['head'].fontRevision
-  bad_ttFonts[1]['head'].fontRevision = version + 1
-
-  print('Test WARN with fonts that diverge on the fontRevision field value.')
-  status, message = list(check(bad_ttFonts))[-1]
-  assert status == WARN
-
-
-def test_check_015():
-  """ Font has post table version 2 ? """
-  from fontbakery.specifications.post import com_google_fonts_check_015 as check
-
-  print('Test PASS with good font.')
-  # our reference Mada family is know to be good here.
-  ttFont = TTFont("data/test/mada/Mada-Regular.ttf")
-  status, message = list(check(ttFont))[-1]
-  assert status == PASS
-
-  # modify the post table version
-  ttFont['post'].formatType = 3
-
-  print('Test FAIL with fonts that diverge on the fontRevision field value.')
-  status, message = list(check(ttFont))[-1]
-  assert status == FAIL
-
-
 def test_check_016():
   """ Checking OS/2 fsType """
   from fontbakery.specifications.googlefonts import com_google_fonts_check_016 as check
@@ -470,17 +335,6 @@ def test_check_016():
   status, message = list(check(ttFont))[-1]
   assert status == FAIL
 
-# DEPRECATED CHECK:
-# com.google.fonts/check/017 - "Assure valid format for the main entries in the name table."
-#
-# REPLACED BY:
-# com.google.fonts/check/156 - "Font has all mandatory 'name' table entries ?"
-# com.google.fonts/check/157 - "Check name table: FONT_FAMILY_NAME entries."
-# com.google.fonts/check/158 - "Check name table: FONT_SUBFAMILY_NAME entries."
-# com.google.fonts/check/159 - "Check name table: FULL_FONT_NAME entries."
-# com.google.fonts/check/160 - "Check name table: POSTSCRIPT_NAME entries."
-# com.google.fonts/check/161 - "Check name table: TYPOGRAPHIC_FAMILY_NAME entries."
-# com.google.fonts/check/162 - "Check name table: TYPOGRAPHIC_SUBFAMILY_NAME entries."
 
 def test_check_018():
   """ Checking OS/2 achVendID """
@@ -532,34 +386,6 @@ def test_check_019():
   status, message = list(check(ttFont))[-1]
   assert status == PASS
 
-
-def test_check_020():
-  """ Checking OS/2 usWeightClass. """
-  from fontbakery.specifications.os2 import com_google_fonts_check_020 as check
-  from fontbakery.specifications.googlefonts import style, expected_os2_weight
-
-  exp_os2_weight = lambda font: expected_os2_weight(style(font))
-  print('Test FAIL with a bad font...')
-  # Our reference Mada Regular is know to be bad here.
-  font = "data/test/mada/Mada-Regular.ttf"
-  status, message = list(check(TTFont(font), exp_os2_weight(font)))[-1]
-  assert status == FAIL
-
-  print('Test PASS with a good font...')
-  # All fonts in our reference Cabin family are know to be good here.
-  for font in cabin_fonts:
-    status, message = list(check(TTFont(font), exp_os2_weight(font)))[-1]
-    assert status == PASS
-
-# DEPRECATED CHECKS:                                             | REPLACED BY:
-# com.google.fonts/check/??? - "Checking macStyle BOLD bit"       | com.google.fonts/check/131 - "Checking head.macStyle value"
-# com.google.fonts/check/021 - "Checking fsSelection REGULAR bit" | com.google.fonts/check/129 - "Checking OS/2.fsSelection value"
-# com.google.fonts/check/022 - "italicAngle <= 0 ?"               | com.google.fonts/check/130 - "Checking post.italicAngle value"
-# com.google.fonts/check/023 - "italicAngle is < 20 degrees ?"    | com.google.fonts/check/130 - "Checking post.italicAngle value"
-# com.google.fonts/check/024 - "italicAngle matches font style ?" | com.google.fonts/check/130 - "Checking post.italicAngle value"
-# com.google.fonts/check/025 - "Checking fsSelection ITALIC bit"  | com.google.fonts/check/129 - "Checking OS/2.fsSelection value"
-# com.google.fonts/check/026 - "Checking macStyle ITALIC bit"     | com.google.fonts/check/131 - "Checking head.macStyle value"
-# com.google.fonts/check/027 - "Checking fsSelection BOLD bit"    | com.google.fonts/check/129 - "Checking OS/2.fsSelection value"
 
 def test_check_028():
   """ Check font project has a license. """
@@ -613,525 +439,7 @@ def test_check_029(mada_ttFonts):
     assert status == FAIL and message.code == 'missing'
 
 # TODO: test_check_030
-
-def test_check_031():
-  """ Description strings in the name table
-      must not contain copyright info.
-  """
-  from fontbakery.specifications.name import com_google_fonts_check_031 as check
-  from fontbakery.constants import NAMEID_DESCRIPTION
-
-  print('Test PASS with a good font...')
-  # Our reference Mada Regular is know to be good here.
-  ttFont = TTFont("data/test/mada/Mada-Regular.ttf")
-  status, message = list(check(ttFont))[-1]
-  assert status == PASS
-
-  # here we add a "Copyright" string to a NAMEID_DESCRIPTION
-  for i, name in enumerate(ttFont['name'].names):
-    if name.nameID == NAMEID_DESCRIPTION:
-      ttFont['name'].names[i].string = "Copyright".encode(name.getEncoding())
-
-  print('Test FAIL with a bad font...')
-  status, message = list(check(ttFont))[-1]
-  assert status == FAIL
-
-
-def test_check_032():
-  """ Description strings in the name table
-      must not exceed 100 characters.
-  """
-  from fontbakery.specifications.name import com_google_fonts_check_032 as check
-  from fontbakery.constants import NAMEID_DESCRIPTION
-
-  print('Test PASS with a good font...')
-  # Our reference Mada Regular is know to be good here.
-  ttFont = TTFont("data/test/mada/Mada-Regular.ttf")
-  status, message = list(check(ttFont))[-1]
-  assert status == PASS
-
-  # Here we add strings to NAMEID_DESCRIPTION with exactly 100 chars,
-  # so it should still PASS:
-  for i, name in enumerate(ttFont['name'].names):
-    if name.nameID == NAMEID_DESCRIPTION:
-      ttFont['name'].names[i].string = ('a' * 100).encode(name.getEncoding())
-
-  print('Test PASS with a 100 char string...')
-  status, message = list(check(ttFont))[-1]
-  assert status == PASS
-
-  # And here we make the strings longer than 100 chars in order to FAIL the check:
-  for i, name in enumerate(ttFont['name'].names):
-    if name.nameID == NAMEID_DESCRIPTION:
-      ttFont['name'].names[i].string = ('a' * 101).encode(name.getEncoding())
-
-  print('Test FAIL with a bad font...')
-  status, message = list(check(ttFont))[-1]
-  assert status == FAIL
-
-
-# If we ever reuse this helper function,
-# then move it into fontbakery.utils:
-def results_contain(results, expected_status, expected_code):
-  for status, message in results:
-    if status == expected_status and message.code == expected_code:
-      return True
-  # else
-  return False
-
-
-def test_check_033():
-  """ Checking correctness of monospaced metadata. """
-  from fontbakery.specifications.name import com_google_fonts_check_033 as check
-  from fontbakery.specifications.shared_conditions import monospace_stats
-  from fontbakery.constants import (PANOSE_PROPORTION__ANY,
-                                    PANOSE_PROPORTION__NO_FIT,
-                                    PANOSE_PROPORTION__OLD_STYLE,
-                                    PANOSE_PROPORTION__MODERN,
-                                    PANOSE_PROPORTION__EVEN_WIDTH,
-                                    PANOSE_PROPORTION__EXTENDED,
-                                    PANOSE_PROPORTION__CONDENSED,
-                                    PANOSE_PROPORTION__VERY_EXTENDED,
-                                    PANOSE_PROPORTION__VERY_CONDENSED,
-                                    PANOSE_PROPORTION__MONOSPACED,
-                                    IS_FIXED_WIDTH__MONOSPACED,
-                                    IS_FIXED_WIDTH__NOT_MONOSPACED)
-
-  # This check has a large number of code-paths
-  # We'll make sure to test them all here.
-  #
-  # --------------------------------------------
-  # Starting with non-monospaced code-paths:
-  # --------------------------------------------
-
-  print('Test PASS with a good non-monospace font...')
-  # Our reference Mada Regular is a non-monospace font
-  # know to have good metadata for this check.
-  ttFont = TTFont("data/test/mada/Mada-Regular.ttf")
-  stats = monospace_stats(ttFont)
-  status, message = list(check(ttFont, stats))[-1]
-  assert status == PASS and message.code == "good"
-
-  # We'll mark it as monospaced on the post table and make sure it fails:
-  print('Test FAIL with a non-monospaced font with bad post.isFixedPitch value ...')
-  ttFont["post"].isFixedPitch = IS_FIXED_WIDTH__MONOSPACED
-  status, message = list(check(ttFont, stats))[-1]
-  assert status == FAIL and message.code == "bad-post-isFixedPitch"
-
-  # restore good value:
-  ttFont["post"].isFixedPitch = IS_FIXED_WIDTH__NOT_MONOSPACED
-
-  # Now we mark it as monospaced on the OS/2 and it should also fail:
-  print('Test FAIL with a non-monospaced font with bad OS/2.panose.bProportion value (MONOSPACED) ...')
-  ttFont["OS/2"].panose.bProportion = PANOSE_PROPORTION__MONOSPACED
-  status, message = list(check(ttFont, stats))[-1]
-  assert status == FAIL and message.code == "bad-panose-proportion"
-
-  # --------------------------------------------
-  # And now we test the monospaced code-paths:
-  # --------------------------------------------
-
-  print('Test PASS with a good monospaced font...')
-  # Our reference OverpassMono Regular is know to be
-  # a monospaced font with good metadata here.
-  ttFont = TTFont("data/test/overpassmono/OverpassMono-Regular.ttf")
-  stats = monospace_stats(ttFont)
-  status, message = list(check(ttFont, stats))[-1]
-  # WARN is emitted when there's at least one outlier.
-  # I don't see a good reason to be picky and also test that one separately here...
-  assert (status == WARN and message.code == "mono-outliers") or \
-         (status == PASS and message.code == "mono-good")
-
-  # Let's incorrectly mark it as a non-monospaced on the post table and it should fail:
-  print('Test FAIL with a monospaced font with bad post.isFixedPitch value ...')
-  ttFont["post"].isFixedPitch = IS_FIXED_WIDTH__NOT_MONOSPACED
-  # here we search for the expected FAIL among all results
-  # instead of simply looking at the last one
-  # because we may also get an outliers WARN in some cases:
-  results = list(check(ttFont, stats))
-  assert results_contain(results, FAIL, "mono-bad-post-isFixedPitch")
-
-  # There are several bad panose proportion values for a monospaced font.
-  # Only PANOSE_PROPORTION__MONOSPACED would be valid.
-  # So we'll try all the bad ones here to make sure all of them emit a FAIL:
-  bad_monospaced_panose_values = [
-    PANOSE_PROPORTION__ANY,
-    PANOSE_PROPORTION__NO_FIT,
-    PANOSE_PROPORTION__OLD_STYLE,
-    PANOSE_PROPORTION__MODERN,
-    PANOSE_PROPORTION__EVEN_WIDTH,
-    PANOSE_PROPORTION__EXTENDED,
-    PANOSE_PROPORTION__CONDENSED,
-    PANOSE_PROPORTION__VERY_EXTENDED,
-    PANOSE_PROPORTION__VERY_CONDENSED,
-  ]
-  good_value = ttFont["OS/2"].panose.bProportion
-  for bad_value in bad_monospaced_panose_values:
-    print('Test FAIL with a monospaced font with bad OS/2.panose.bProportion value ({}) ...'.format(bad_value))
-    ttFont["OS/2"].panose.bProportion = bad_value
-    # again, we search the expected FAIL because we may algo get an outliers WARN here:
-    results = list(check(ttFont, stats))
-    assert results_contain(results, FAIL, "mono-bad-panose-proportion")
-
-
-# TODO: test_check_034
-# TODO: test_check_035
-# TODO: test_check_036
-
-def test_check_037():
-  """ MS Font Validator checks """
-  from fontbakery.specifications.general import com_google_fonts_check_037 as check
-
-  font = "data/test/mada/Mada-Regular.ttf"
-  RASTER_EXCEPTION_MESSAGE = ("MS-FonVal: An exception occurred"
-                              " during rasterization testing")
-  # we want to run all FValidator checks only once,
-  # so here we cache all results:
-  fval_results = list(check(font))
-
-  # Then we make sure that the freetype backend we're using
-  # supports the hinting instructions validation checks,
-  # which are refered to as "rasterization testing":
-  # (See also: https://github.com/googlefonts/fontbakery/issues/1524)
-  for status, message in fval_results:
-    assert RASTER_EXCEPTION_MESSAGE not in message
-
-  # and finaly, we make sure that there wasn't an ERROR
-  # which would mean FontValidator is not properly installed:
-  assert status != ERROR
-
-# TODO: test_check_038
-# TODO: test_check_039
-
-def test_check_040(mada_ttFonts):
-  """ Checking OS/2 usWinAscent & usWinDescent. """
-  from fontbakery.specifications.os2 import com_google_fonts_check_040 as check
-  from fontbakery.specifications.shared_conditions import vmetrics
-
-  # Our reference Mada Regular is know to be bad here.
-  vm = vmetrics(mada_ttFonts)
-  ttFont = TTFont("data/test/mada/Mada-Regular.ttf")
-
-  # But we fix it first to test the PASS code-path:
-  print('Test PASS with a good font...')
-  ttFont['OS/2'].usWinAscent = vm['ymax']
-  ttFont['OS/2'].usWinDescent = abs(vm['ymin'])
-  status, message = list(check(ttFont, vm))[-1]
-  assert status == PASS
-
-  # Then we break it:
-  print('Test FAIL with a bad OS/2.usWinAscent...')
-  ttFont['OS/2'].usWinAscent = vm['ymax'] - 1
-  ttFont['OS/2'].usWinDescent = abs(vm['ymin'])
-  status, message = list(check(ttFont, vm))[-1]
-  assert status == FAIL and message.code == "ascent"
-
-  print('Test FAIL with a bad OS/2.usWinDescent...')
-  ttFont['OS/2'].usWinAscent = vm['ymax']
-  ttFont['OS/2'].usWinDescent = abs(vm['ymin']) - 1
-  status, message = list(check(ttFont, vm))[-1]
-  assert status == FAIL and message.code == "descent"
-
-
-def test_check_041():
-  """ Checking Vertical Metric Linegaps. """
-  from fontbakery.specifications.hhea import com_google_fonts_check_041 as check
-
-  print('Test FAIL with non-zero hhea.lineGap...')
-  # Our reference Mada Regular is know to be bad here.
-  ttFont = TTFont("data/test/mada/Mada-Regular.ttf")
-
-  # But just to be sure, we first explicitely set
-  # the values we're checking for:
-  ttFont['hhea'].lineGap = 1
-  ttFont['OS/2'].sTypoLineGap = 0
-  status, message = list(check(ttFont))[-1]
-  assert status == WARN and message.code == "hhea"
-
-  # Then we run the check with a non-zero OS/2.sTypoLineGap:
-  ttFont['hhea'].lineGap = 0
-  ttFont['OS/2'].sTypoLineGap = 1
-  status, message = list(check(ttFont))[-1]
-  assert status == WARN and message.code == "OS/2"
-
-  # And finaly we fix it by making both values equal to zero:
-  ttFont['hhea'].lineGap = 0
-  ttFont['OS/2'].sTypoLineGap = 0
-  status, message = list(check(ttFont))[-1]
-  assert status == PASS
-
-
-def test_check_042(mada_ttFonts):
-  """ Checking OS/2 Metrics match hhea Metrics. """
-  from fontbakery.specifications.os2 import com_google_fonts_check_042 as check
-
-  # Our reference Mada Regular is know to be good here.
-  ttFont = TTFont("data/test/mada/Mada-Regular.ttf")
-
-  print("Test PASS with a good font...")
-  status, message = list(check(ttFont))[-1]
-  assert status == PASS
-
-  # Now we break it:
-  print('Test FAIL with a bad OS/2.sTypoAscender font...')
-  correct = ttFont['hhea'].ascent
-  ttFont['OS/2'].sTypoAscender = correct + 1
-  status, message = list(check(ttFont))[-1]
-  assert status == FAIL and message.code == "ascender"
-
-  # Restore good value:
-  ttFont['OS/2'].sTypoAscender = correct
-
-  # And break it again, now on sTypoDescender value:
-  print('Test FAIL with a bad OS/2.sTypoDescender font...')
-  correct = ttFont['hhea'].descent
-  ttFont['OS/2'].sTypoDescender = correct + 1
-  status, message = list(check(ttFont))[-1]
-  assert status == FAIL and message.code == "descender"
-
-
-def test_check_043():
-  """ Checking unitsPerEm value is reasonable. """
-  from fontbakery.specifications.head import com_google_fonts_check_043 as check
-
-  # In this test we'll forge several known-good and known-bad values.
-  # We'll use Mada Regular to start with:
-  ttFont = TTFont("data/test/mada/Mada-Regular.ttf")
-
-  for good_value in [1000, 16, 32, 64, 128, 256,
-                     512, 1024, 2048, 4096, 8192, 16384]:
-    print("Test PASS with a good value of unitsPerEm = {} ...".format(good_value))
-    ttFont['head'].unitsPerEm = good_value
-    status, message = list(check(ttFont))[-1]
-    assert status == PASS
-
-  # These are arbitrarily chosen bad values:
-  for bad_value in [0, 1, 2, 4, 8, 10, 100, 10000, 32768]:
-    print("Test FAIL with a bad value of unitsPerEm = {} ...".format(bad_value))
-    ttFont['head'].unitsPerEm = bad_value
-    status, message = list(check(ttFont))[-1]
-    assert status == FAIL
-
-# TODO: test_check_044
-
-def test_check_045():
-  """ Does the font have a DSIG table ? """
-  from fontbakery.specifications.dsig import com_google_fonts_check_045 as check
-
-  # Our reference Cabin Regular font is good (theres a DSIG table declared):
-  ttFont = TTFont("data/test/cabin/Cabin-Regular.ttf")
-
-  # So it must PASS the check:
-  print ("Test PASS with a good font...")
-  status, message = list(check(ttFont))[-1]
-  assert status == PASS
-
-  # Then we remove the DSIG table so that we get a FAIL:
-  print ("Test FAIL with a font lacking a DSIG table...")
-  del ttFont['DSIG']
-  status, message = list(check(ttFont))[-1]
-  assert status == FAIL
-
-# TODO: test_check_046
-
-def test_check_047():
-  """ Font contains glyphs for whitespace characters ? """
-  from fontbakery.specifications.general import com_google_fonts_check_047 as check
-  from fontbakery.specifications.shared_conditions import missing_whitespace_chars
-
-  # Our reference Mada Regular font is good here:
-  ttFont = TTFont("data/test/mada/Mada-Regular.ttf")
-  missing = missing_whitespace_chars(ttFont)
-
-  # So it must PASS the check:
-  print ("Test PASS with a good font...")
-  status, message = list(check(ttFont, missing))[-1]
-  assert status == PASS
-
-  # Then we remove the nbsp char (0x00A0) so that we get a FAIL:
-  print ("Test FAIL with a font lacking a nbsp (0x00A0)...")
-  for table in ttFont['cmap'].tables:
-    if 0x00A0 in table.cmap:
-      del table.cmap[0x00A0]
-
-  missing = missing_whitespace_chars(ttFont)
-  status, message = list(check(ttFont, missing))[-1]
-  assert status == FAIL
-
-  # restore original Mada Regular font:
-  ttFont = TTFont("data/test/mada/Mada-Regular.ttf")
-
-  # And finally remove the space character (0x0020) to get another FAIL:
-  print ("Test FAIL with a font lacking a space (0x0020)...")
-  for table in ttFont['cmap'].tables:
-    if 0x00A0 in table.cmap:
-      del table.cmap[0x00A0]
-
-  missing = missing_whitespace_chars(ttFont)
-  status, message = list(check(ttFont, missing))[-1]
-  assert status == FAIL
-
-
-def test_check_048():
-  """ Font has **proper** whitespace glyph names ? """
-  from fontbakery.specifications.general import com_google_fonts_check_048 as check
-
-  def deleteGlyphEncodings(font, cp):
-    """ This routine is used on to introduce errors
-        in a given font by removing specific entries
-        in the cmap tables.
-    """
-    for subtable in font['cmap'].tables:
-      if subtable.isUnicode():
-        for codepoint, name in subtable.cmap.items():
-          if codepoint == cp:
-            del subtable.cmap[codepoint]
-
-  # Our reference Mada Regular font is good here:
-  ttFont = TTFont("data/test/mada/Mada-Regular.ttf")
-
-  # So it must PASS the check:
-  print ("Test PASS with a good font...")
-  status, message = list(check(ttFont))[-1]
-  assert status == PASS
-
-  print ("Test SKIP with post.formatType == 3.0 ...")
-  value = ttFont["post"].formatType
-  ttFont["post"].formatType = 3.0
-  status, message = list(check(ttFont))[-1]
-  assert status == SKIP
-  # and restore good value:
-  ttFont["post"].formatType = value
-
-  print ("Test FAIL with bad glyph name for char 0x0020 ...")
-  deleteGlyphEncodings(ttFont, 0x0020)
-  status, message = list(check(ttFont))[-1]
-  assert status == FAIL and message.code == "bad20"
-
-  # restore the original font object in preparation for the next test-case:
-  ttFont = TTFont("data/test/mada/Mada-Regular.ttf")
-
-  print ("Test FAIL with bad glyph name for char 0x00A0 ...")
-  deleteGlyphEncodings(ttFont, 0x00A0)
-  status, message = list(check(ttFont))[-1]
-  assert status == FAIL and message.code == "badA0"
-
-# TODO: test_check_049
 # TODO: test_check_050 (the original check itself has unclear semantics, so that needs to be reviewed first)
-
-# DEPRECATED:
-# com.google.fonts/check/051 - "Checking with pyfontaine"
-#
-# Replaced by:
-# com.google.fonts/check/132 - "Checking Google Cyrillic Historical glyph coverage"
-# com.google.fonts/check/133 - "Checking Google Cyrillic Plus glyph coverage"
-# com.google.fonts/check/134 - "Checking Google Cyrillic Plus (Localized Forms) glyph coverage"
-# com.google.fonts/check/135 - "Checking Google Cyrillic Pro glyph coverage"
-# com.google.fonts/check/136 - "Checking Google Greek Ancient Musical Symbols glyph coverage"
-# com.google.fonts/check/137 - "Checking Google Greek Archaic glyph coverage"
-# com.google.fonts/check/138 - "Checking Google Greek Coptic glyph coverage"
-# com.google.fonts/check/139 - "Checking Google Greek Core glyph coverage"
-# com.google.fonts/check/140 - "Checking Google Greek Expert glyph coverage"
-# com.google.fonts/check/141 - "Checking Google Greek Plus glyph coverage"
-# com.google.fonts/check/142 - "Checking Google Greek Pro glyph coverage"
-# com.google.fonts/check/143 - "Checking Google Latin Core glyph coverage"
-# com.google.fonts/check/144 - "Checking Google Latin Expert glyph coverage"
-# com.google.fonts/check/145 - "Checking Google Latin Plus glyph coverage"
-# com.google.fonts/check/146 - "Checking Google Latin Plus (Optional Glyphs) glyph coverage"
-# com.google.fonts/check/147 - "Checking Google Latin Pro glyph coverage"
-# com.google.fonts/check/148 - "Checking Google Latin Pro (Optional Glyphs) glyph coverage"
-
-
-def test_check_052():
-  """ Font contains all required tables ? """
-  from fontbakery.specifications.general import com_google_fonts_check_052 as check
-
-  required_tables = ["cmap", "head", "hhea", "hmtx",
-                     "maxp", "name", "OS/2", "post"]
-  optional_tables = ["cvt ", "fpgm", "loca", "prep",
-                     "VORG", "EBDT", "EBLC", "EBSC",
-                     "BASE", "GPOS", "GSUB", "JSTF",
-                     "DSIG", "gasp", "hdmx", "kern",
-                     "LTSH", "PCLT", "VDMX", "vhea",
-                     "vmtx"]
-  # Our reference Mada Regular font is good here
-  ttFont = TTFont("data/test/mada/Mada-Regular.ttf")
-
-  # So it must PASS the check:
-  print ("Test PASS with a good font...")
-  status, message = list(check(ttFont))[-1]
-  assert status == PASS
-
-  # Now we test the special cases for variable fonts:
-
-  print ("Test FAIL with fvar but no STAT...")
-  # Note: A TTF with an fvar table but no STAT table
-  #       is probably a GX font. For now we're OK with
-  #       rejecting those by emitting a FAIL in this case.
-  #
-  # TODO: Maybe we could someday emit a more explicit
-  #       message to the users regarding that...
-  ttFont.reader.tables["fvar"] = "foo"
-  status, message = list(check(ttFont))[-1]
-  assert status == FAIL
-
-  print ("Test PASS with STAT on a non-variable font...")
-  del ttFont.reader.tables["fvar"]
-  ttFont.reader.tables["STAT"] = "foo"
-  status, message = list(check(ttFont))[-1]
-  assert status == PASS
-
-  # and finally remove what we've just added:
-  del ttFont.reader.tables["STAT"]
-
-  # Now we remove required tables one-by-one to validate the FAIL code-path:
-  for required in required_tables:
-    print ("Test FAIL with missing mandatory table {} ...".format(required))
-    ttFont = TTFont("data/test/mada/Mada-Regular.ttf")
-    if required in ttFont.reader.tables:
-      del ttFont.reader.tables[required]
-    status, message = list(check(ttFont))[-1]
-    assert status == FAIL
-
-  # Then, in preparation for the next step, we make sure
-  # there's no optional table (by removing them all):
-  for optional in optional_tables:
-    if optional in ttFont.reader.tables:
-      del ttFont.reader.tables[optional]
-
-  # Then re-insert them one by one to validate the INFO code-path:
-  for optional in optional_tables:
-    print ("Test INFO with optional table {} ...".format(required))
-    ttFont.reader.tables[optional] = "foo"
-    # and ensure that the second to last logged message is an
-    # INFO status informing the user about it:
-    status, message = list(check(ttFont))[-2]
-    assert status == INFO
-    # remove the one we've just inserted before trying the next one:
-    del ttFont.reader.tables[optional]
-
-
-def test_check_053():
-  """ Are there unwanted tables ? """
-  from fontbakery.specifications.general import com_google_fonts_check_053 as check
-
-  unwanted_tables = ["FFTM", "TTFA", "prop"]
-  # Our reference Mada Regular font is good here:
-  ttFont = TTFont("data/test/mada/Mada-Regular.ttf")
-
-  # So it must PASS the check:
-  print ("Test PASS with a good font...")
-  status, message = list(check(ttFont))[-1]
-  assert status == PASS
-
-  # We now add unwanted tables one-by-one to validate the FAIL code-path:
-  for unwanted in unwanted_tables:
-    print ("Test FAIL with unwanted table {} ...".format(unwanted))
-    ttFont = TTFont("data/test/mada/Mada-Regular.ttf")
-    ttFont.reader.tables[unwanted] = "foo"
-    status, message = list(check(ttFont))[-1]
-    assert status == FAIL
-
 # TODO: test_check_054
 
 def test_check_055():
@@ -1166,30 +474,6 @@ def test_check_055():
 
 # TODO: test_check_056
 
-def test_check_057():
-  """ Name table entries should not contain line-breaks. """
-  from fontbakery.specifications.name import com_google_fonts_check_057 as check
-
-  # Our reference Mada Regular font is good here:
-  ttFont = TTFont("data/test/mada/Mada-Regular.ttf")
-
-  # So it must PASS the check:
-  print ("Test PASS with a good font...")
-  status, message = list(check(ttFont))[-1]
-  assert status == PASS
-
-  print ("Test FAIL with name entries containing a linebreak...")
-  for i in range(len(ttFont["name"].names)):
-    ttFont = TTFont("data/test/mada/Mada-Regular.ttf")
-    encoding = ttFont["name"].names[i].getEncoding()
-    ttFont["name"].names[i].string = "bad\nstring".encode(encoding)
-    status, message = list(check(ttFont))[-1]
-    assert status == FAIL
-
-# TODO: test_check_058
-# TODO: test_check_059
-# TODO: test_check_060
-
 def test_check_061():
   """ EPAR table present in font ? """
   from fontbakery.specifications.googlefonts import com_google_fonts_check_061 as check
@@ -1210,129 +494,6 @@ def test_check_061():
   assert status == PASS
 
 # TODO: test_check_062
-
-def test_check_063():
-  """ Does GPOS table have kerning information ? """
-  from fontbakery.specifications.gpos import com_google_fonts_check_063 as check
-
-  # Our reference Mada Regular is known to have kerning-info
-  # exclusively on an extension subtable
-  # (lookup type = 9 / ext-type = 2):
-  ttFont = TTFont("data/test/mada/Mada-Regular.ttf")
-
-  # So it must PASS the check:
-  print ("Test PASS with a font that has got kerning info...")
-  status, message = list(check(ttFont))[-1]
-  assert status == PASS
-
-  # delete all Pair Adjustment lookups:
-  while True:
-    found = False
-    for l, lookup in enumerate(ttFont["GPOS"].table.LookupList.Lookup):
-      #if lookup.LookupType == 2:  # type 2 = Pair Adjustment
-      #  del ttFont["GPOS"].table.LookupList.Lookup[l]
-      #  found = True
-      if lookup.LookupType == 9:  # type 9 = Extension subtable
-        for e, ext in enumerate(lookup.SubTable):
-          if ext.ExtensionLookupType == 2:  # type 2 = Pair Adjustment
-            del ttFont["GPOS"].table.LookupList.Lookup[l].SubTable[e]
-            found = True
-    if not found:
-      break
-
-  print ("Test WARN with a font lacking kerning info...")
-  status, message = list(check(ttFont))[-1]
-  assert status == WARN
-
-  # setup a fake type=2 Pair Adjustment lookup
-  ttFont["GPOS"].table.LookupList.Lookup[0].LookupType = 2
-  # and make sure the check emits a PASS result:
-  print ("Test PASS with kerning info on a type=2 lookup...")
-  status, message = list(check(ttFont))[-1]
-  assert status == PASS
-
-  # remove the GPOS table and make sure to get a WARN:
-  del ttFont["GPOS"]
-  print ("Test WARN with a font lacking a GPOS table...")
-  status, message = list(check(ttFont))[-1]
-  assert status == WARN
-
-
-def test_check_064():
-  """ Is there a caret position declared for every ligature ? """
-  from fontbakery.specifications.gdef import com_google_fonts_check_064 as check
-  from fontbakery.specifications.shared_conditions import ligatures
-
-  # TODO: We currently lack a reference family that PASSes this check!
-
-  # Our reference Mada Medium is known to be bad
-  ttFont = TTFont("data/test/mada/Mada-Medium.ttf")
-  lig = ligatures(ttFont)
-
-  # So it must emit a WARN:
-  print ("Test WARN with a bad font...")
-  status, message = list(check(ttFont, lig))[-1]
-  assert status == WARN and message.code == "lacks-caret-pos"
-
-  # And FamilySans Regular is known to be bad
-  ttFont = TTFont("data/test/familysans/FamilySans-Regular.ttf")
-  lig = ligatures(ttFont)
-
-  # So it must emit a WARN:
-  print ("Test WARN with a bad font...")
-  status, message = list(check(ttFont, lig))[-1]
-  assert status == WARN and message.code == "GDEF-missing"
-
-  # TODO: test the WARN "incomplete-caret-pos-data" codepath
-
-
-def test_check_065():
-  """ Is there kerning info for non-ligated sequences ? """
-  from fontbakery.specifications.gpos import (com_google_fonts_check_065 as check,
-                                                     has_kerning_info)
-  from fontbakery.specifications.shared_conditions import ligatures
-  # Our reference Mada Medium is known to be good
-  ttFont = TTFont("data/test/mada/Mada-Medium.ttf")
-  lig = ligatures(ttFont)
-  has_kinfo = has_kerning_info(ttFont)
-
-  # So it must PASS the check:
-  print ("Test PASS with a good font...")
-  status, message = list(check(ttFont, lig, has_kinfo))[-1]
-  assert status == PASS
-
-  # And Merriweather Regular is known to be bad
-  ttFont = TTFont("data/test/merriweather/Merriweather-Regular.ttf")
-  lig = ligatures(ttFont)
-  has_kinfo = has_kerning_info(ttFont)
-
-  # So the check must emit a WARN in this testcase:
-  print ("Test WARN with a bad font...")
-  status, message = list(check(ttFont, lig, has_kinfo))[-1]
-  assert status == WARN and message.code == "lacks-kern-info"
-
-
-def test_check_066():
-  """ Is there a "kern" table declared in the font ? """
-  from fontbakery.specifications.kern import com_google_fonts_check_066 as check
-
-  # Our reference Mada Regular is known to be good
-  # (does not have a 'kern' table):
-  ttFont = TTFont("data/test/mada/Mada-Regular.ttf")
-
-  # So it must PASS the check:
-  print ("Test PASS with a font without a 'kern' table...")
-  status, message = list(check(ttFont))[-1]
-  assert status == PASS
-
-  # add a fake 'kern' table:
-  ttFont["kern"] = "foo"
-
-  # and make sure the check emits an INFO message:
-  print ("Test INFO with a font containing a 'kern' table...")
-  status, message = list(check(ttFont))[-1]
-  assert status == INFO
-
 
 def test_check_067():
   """ Make sure family name does not begin with a digit. """
@@ -1358,47 +519,6 @@ def test_check_067():
   assert status == FAIL
 
 
-def test_check_068():
-  """ Does full font name begin with the font family name ? """
-  from fontbakery.specifications.name import com_google_fonts_check_068 as check
-  from fontbakery.constants import (NAMEID_FULL_FONT_NAME,
-                                    NAMEID_FONT_FAMILY_NAME)
-  # Our reference Mada Regular is known to be good
-  ttFont = TTFont("data/test/mada/Mada-Regular.ttf")
-
-  # So it must PASS the check:
-  print ("Test PASS with a good font...")
-  status, message = list(check(ttFont))[-1]
-  assert status == PASS
-
-  # alter the full-font-name prepending a bad prefix:
-  for i, name in enumerate(ttFont["name"].names):
-    if name.nameID == NAMEID_FULL_FONT_NAME:
-      ttFont["name"].names[i].string = "bad-prefix".encode(name.getEncoding())
-
-  # and make sure the check FAILs:
-  print ("Test FAIL with a font in which the family name begins with a digit...")
-  status, message = list(check(ttFont))[-1]
-  assert status == FAIL and message.code == "does-not"
-
-  print ("Test FAIL with no FULL_FONT_NAME entries...")
-  ttFont = TTFont("data/test/mada/Mada-Regular.ttf")
-  for i, name in enumerate(ttFont["name"].names):
-    if name.nameID == NAMEID_FULL_FONT_NAME:
-      del ttFont["name"].names[i]
-  status, message = list(check(ttFont))[-1]
-  assert status == FAIL and message.code == "no-full-font-name"
-
-  print ("Test FAIL with no FONT_FAMILY_NAME entries...")
-  ttFont = TTFont("data/test/mada/Mada-Regular.ttf")
-  for i, name in enumerate(ttFont["name"].names):
-    if name.nameID == NAMEID_FONT_FAMILY_NAME:
-      del ttFont["name"].names[i]
-  status, message = list(check(ttFont))[-1]
-  assert status == FAIL and message.code == "no-font-family-name"
-
-# TODO: test/069
-
 def test_check_070():
   """ Font has all expected currency sign characters ? """
   from fontbakery.specifications.googlefonts import com_google_fonts_check_070 as check
@@ -1419,125 +539,7 @@ def test_check_070():
   status, message = list(check(ttFont))[-1]
   assert status == FAIL
 
-
-def assert_name_table_check_result(ttFont, index, name, check, value, expected_result):
-  backup = name.string
-  # set value
-  ttFont["name"].names[index].string = value.encode(name.getEncoding())
-  # run check
-  status, message = list(check(ttFont))[-1]
-  # restore value
-  ttFont["name"].names[index].string = backup
-  assert status == expected_result
-
-
-def test_check_071():
-  """ Font follows the family naming recommendations ? """
-  from fontbakery.specifications.name import com_google_fonts_check_071 as check
-  from fontbakery.constants import (NAMEID_POSTSCRIPT_NAME,
-                                    NAMEID_FULL_FONT_NAME,
-                                    NAMEID_FONT_FAMILY_NAME,
-                                    NAMEID_FONT_SUBFAMILY_NAME,
-                                    NAMEID_TYPOGRAPHIC_FAMILY_NAME,
-                                    NAMEID_TYPOGRAPHIC_SUBFAMILY_NAME)
-  # Our reference Mada Medium is known to be good
-  ttFont = TTFont("data/test/mada/Mada-Medium.ttf")
-
-  # So it must PASS the check:
-  print ("Test PASS with a good font...")
-  status, message = list(check(ttFont))[-1]
-  assert status == PASS
-
-  # We'll test rule violations in all entries one-by-one
-  for index, name in enumerate(ttFont["name"].names):
-    # and we'll test all INFO/PASS code-paths for each of the rules:
-    def name_test(value, expected):
-      assert_name_table_check_result(ttFont, index, name, check, value, expected) #pylint: disable=cell-var-from-loop
-
-    if name.nameID == NAMEID_POSTSCRIPT_NAME:
-      print ("== NAMEID_POST_SCRIPT_NAME ==")
-
-      print ("Test INFO: May contain only a-zA-Z0-9 characters and an hyphen...")
-      # The '@' and '!' chars here are the expected rule violations:
-      name_test("B@zinga!", INFO)
-
-      print ("Test PASS: A name with a single hyphen is OK...")
-      # A single hypen in the name is OK:
-      name_test("Big-Bang", PASS)
-
-      print ("Test INFO: May not contain more than a single hyphen...")
-      # The second hyphen char here is the expected rule violation:
-      name_test("Big-Bang-Theory", INFO)
-
-      print ("Test INFO: Exceeds max length (29)...")
-      name_test("A"*30, INFO)
-
-      print ("Test PASS: Does not exceeds max length...")
-      name_test("A"*29, PASS)
-
-    elif name.nameID == NAMEID_FULL_FONT_NAME:
-      print ("== NAMEID_FULL_FONT_NAME ==")
-
-      print ("Test INFO: Exceeds max length (63)...")
-      name_test("A"*64, INFO)
-
-      print ("Test PASS: Does not exceeds max length...")
-      name_test("A"*63, PASS)
-
-    elif name.nameID == NAMEID_FONT_FAMILY_NAME:
-      print ("== NAMEID_FONT_FAMILY_NAME ==")
-
-      print ("Test INFO: Exceeds max length (31)...")
-      name_test("A"*32, INFO)
-
-      print ("Test PASS: Does not exceeds max length...")
-      name_test("A"*31, PASS)
-
-    elif name.nameID == NAMEID_FONT_SUBFAMILY_NAME:
-      print ("== NAMEID_FONT_SUBFAMILY_NAME ==")
-
-      print ("Test INFO: Exceeds max length (31)...")
-      name_test("A"*32, INFO)
-
-      print ("Test PASS: Does not exceeds max length...")
-      name_test("A"*31, PASS)
-
-    elif name.nameID == NAMEID_TYPOGRAPHIC_FAMILY_NAME:
-      print ("== NAMEID_TYPOGRAPHIC_FAMILY_NAME ==")
-
-      print ("Test INFO: Exceeds max length (31)...")
-      name_test("A"*32, INFO)
-
-      print ("Test PASS: Does not exceeds max length...")
-      name_test("A"*31, PASS)
-
-    elif name.nameID == NAMEID_TYPOGRAPHIC_SUBFAMILY_NAME:
-      print ("== NAMEID_FONT_TYPOGRAPHIC_SUBFAMILY_NAME ==")
-
-      print ("Test INFO: Exceeds max length (31)...")
-      name_test("A"*32, INFO)
-
-      print ("Test PASS: Does not exceeds max length...")
-      name_test("A"*31, PASS)
-
-  for w in range(0, 1000, 50):
-    ttFont["OS/2"].usWeightClass = w
-    if w < 250 or w > 900:
-      print ("Test FAIL: Weight out of acceptable range of values (from 250 to 900)...")
-      status, message = list(check(ttFont))[-1]
-      assert status == INFO
-    else:
-      print ("Test PASS: Weight value is between 250 and 900 (including the extreme values)...")
-      status, message = list(check(ttFont))[-1]
-      assert status == PASS
-
-  for w in [251, 275, 325, 425, 775, 825, 899]:
-    ttFont["OS/2"].usWeightClass = w
-    print ("Test FAIL: Weight value is not multiple of 50...")
-    status, message = list(check(ttFont))[-1]
-    assert status == INFO
-
-# TODO: checks 072 to 093
+# FIXME: TODO: checks 072 to 093
 
 def test_check_094():
   """ METADATA.pb font.fullname value matches fullname declared on the name table ? """
@@ -1598,7 +600,6 @@ def test_check_095():
       ttFont["name"].names[i].string = good # restore good value
 
 # TODO: test_check_096
-
 # TODO: test_check_097
 
 MONTSERRAT_RIBBI = [
@@ -1989,34 +990,7 @@ def test_check_112():
     assert status == FAIL
 
 # TODO: test_check_113
-
-# DEPRECATED: com.google.fonts/check/114
-
-# TODO: checks 115 to 131
-
-# DEPRECATED CHECKS:
-# com.google.fonts/check/132 - "Checking Cyrillic Historical glyph coverage."
-# com.google.fonts/check/133 - "Checking Google Cyrillic Plus glyph coverage."
-# com.google.fonts/check/134 - "Checking Google Cyrillic Plus (Localized Forms) glyph coverage."
-# com.google.fonts/check/135 - "Checking Google Cyrillic Pro glyph coverage."
-# com.google.fonts/check/136 - "Checking Google Greek Ancient Musical Symbols glyph coverage."
-# com.google.fonts/check/137 - "Checking Google Greek Archaic glyph coverage."
-# com.google.fonts/check/138 - "Checking Google Greek Coptic glyph coverage."
-# com.google.fonts/check/139 - "Checking Google Greek Core glyph coverage."
-# com.google.fonts/check/140 - "Checking Google Greek Expert glyph coverage."
-# com.google.fonts/check/141 - "Checking Google Greek Plus glyph coverage."
-# com.google.fonts/check/142 - "Checking Google Greek Pro glyph coverage."
-# com.google.fonts/check/143 - "Checking Google Latin Core glyph coverage."
-# com.google.fonts/check/144 - "Checking Google Latin Expert glyph coverage."
-# com.google.fonts/check/145 - "Checking Google Latin Plus glyph coverage."
-# com.google.fonts/check/146 - "Checking Google Latin Plus (Optional Glyphs) glyph coverage."
-# com.google.fonts/check/147 - "Checking Google Latin Pro glyph coverage."
-# com.google.fonts/check/148 - "Checking Google Latin Pro (Optional Glyphs) glyph coverage."
-# com.google.fonts/check/149 - "Checking Google Arabic glyph coverage."
-# com.google.fonts/check/150 - "Checking Google Vietnamese glyph coverage."
-# com.google.fonts/check/151 - "Checking Google Extras glyph coverage."
-
-# TODO: test_check_152
+# FIXME: TODO: checks 115 to 131
 
 def test_check_153(montserrat_ttFonts):
   """Check glyphs contain the recommended contour count"""
@@ -2033,35 +1007,37 @@ def test_check_153(montserrat_ttFonts):
     status, message = list(check(ttFont))[-1]
     assert status == WARN
 
+
 # FIXME! This works fine locally, but crashes on Travis
 # See: https://travis-ci.org/googlefonts/fontbakery/builds/341946880
 # and also: https://github.com/googlefonts/fontbakery/issues/1712
 #
-#def test_check_154(cabin_ttFonts):
-#    """Check glyphs are not missing when compared to version on fonts.google.com"""
-#    from fontbakery.specifications.googlefonts import (com_google_fonts_check_154 as check,
-#                                                       api_gfonts_ttFont,
-#                                                       remote_styles,
-#                                                       metadata)
-#    font = cabin_ttFonts[-1]
-#    print(cabin_ttFonts)
-#    style = font['name'].getName(2, 1, 0, 0)
-#
-#    meta = metadata("data/test/regression/cabin/")
-#    gfonts_remote_styles = remote_styles(meta)
-#    gfont = api_gfonts_ttFont(str(style), gfonts_remote_styles)
-#
-#    # Cabin font hosted on fonts.google.com contains
-#    # all the glyphs for the font in data/test/cabin
-#    status, message = list(check(font, gfont))[-1]
-#    assert status == PASS
-#
-#    # Take A glyph out of font
-#    font['cmap'].getcmap(3, 1).cmap.pop(ord('A'))
-#    font['glyf'].glyphs.pop('A')
-#
-#    status, message = list(check(font, gfont))[-1]
-#    assert status == FAIL
+def disabled_test_check_154(cabin_ttFonts):
+    """Check glyphs are not missing when compared to version on fonts.google.com"""
+    from fontbakery.specifications.googlefonts import (com_google_fonts_check_154 as check,
+                                                       api_gfonts_ttFont,
+                                                       remote_styles,
+                                                       metadata)
+    font = cabin_ttFonts[-1]
+    print(cabin_ttFonts)
+    style = font['name'].getName(2, 1, 0, 0)
+
+    meta = metadata("data/test/regression/cabin/")
+    gfonts_remote_styles = remote_styles(meta)
+    gfont = api_gfonts_ttFont(str(style), gfonts_remote_styles)
+
+    # Cabin font hosted on fonts.google.com contains
+    # all the glyphs for the font in data/test/cabin
+    status, message = list(check(font, gfont))[-1]
+    assert status == PASS
+
+    # Take A glyph out of font
+    font['cmap'].getcmap(3, 1).cmap.pop(ord('A'))
+    font['glyf'].glyphs.pop('A')
+
+    status, message = list(check(font, gfont))[-1]
+    assert status == FAIL
+
 
 def test_check_155():
   """ Copyright field for this font on METADATA.pb matches
@@ -2205,37 +1181,6 @@ def test_check_159():
 # TODO: test_check_161
 # TODO: test_check_162
 
-def test_check_163():
-  """ Check font name is the same as family name. """
-  from fontbakery.specifications.name import com_google_fonts_check_163 as check
-  from fontbakery.constants import (NAMEID_FONT_FAMILY_NAME,
-                                    NAMEID_FONT_SUBFAMILY_NAME)
-  # Our reference Cabin Regular is known to be good
-  ttFont = TTFont("data/test/cabin/Cabin-Regular.ttf")
-
-  # So it must PASS the check:
-  print ("Test PASS with a good font...")
-  status, message = list(check(ttFont))[-1]
-  assert status == PASS
-
-  # Then we emit a WARNing with the long family/style names
-  # that were used as an example on the glyphs tutorial
-  # (at https://glyphsapp.com/tutorials/multiple-masters-part-3-setting-up-instances):
-  for index, name in enumerate(ttFont["name"].names):
-    if name.nameID == NAMEID_FONT_FAMILY_NAME:
-      ttFont["name"].names[index].string = "ImpossibleFamilyNameFont".encode(name.getEncoding())
-      break
-
-  for index, name in enumerate(ttFont["name"].names):
-    if name.nameID == NAMEID_FONT_SUBFAMILY_NAME:
-      ttFont["name"].names[index].string = "WithAVeryLongStyleName".encode(name.getEncoding())
-      break
-
-  print ("Test WARN with a bad font...")
-  status, message = list(check(ttFont))[-1]
-  assert status == WARN
-
-
 def test_check_164():
   """ Length of copyright notice must not exceed 500 characters. """
   from fontbakery.specifications.googlefonts import com_google_fonts_check_164 as check
@@ -2312,216 +1257,6 @@ def test_check_166():
   status, message = list(check(ttFont))[-1]
   assert status == PASS
 
-
-def test_check_167():
-  """ The variable font 'wght' (Weight) axis coordinate
-      must be 400 on the 'Regular' instance. """
-  from fontbakery.specifications.fvar import com_google_fonts_check_167 as check
-  from fontbakery.specifications.shared_conditions import regular_wght_coord
-
-  # Our reference varfont, CabinVFBeta.ttf, has
-  # a bad Regular:wght coordinate
-  ttFont = TTFont("data/test/cabinvfbeta/CabinVFBeta.ttf")
-  regular_weight_coord = regular_wght_coord(ttFont)
-
-  # So it must FAIL the test
-  print('Test FAIL with a bad Regular:wght coordinate...')
-  status, message = list(check(ttFont, regular_weight_coord))[-1]
-  assert status == FAIL
-
-  # We then fix the value:
-  ttFont["fvar"].instances[0].coordinates["wght"] = 400
-  # Note: I know the correct instance index for this hotfix because
-  # I inspected the our reference CabinVF using ttx
-
-  # Then re-read the coord:
-  regular_weight_coord = regular_wght_coord(ttFont)
-
-  # and now this should PASS the test:
-  print('Test PASS with a good Regular:wght coordinate (400)...')
-  status, message = list(check(ttFont, regular_weight_coord))[-1]
-  assert status == PASS
-
-
-def test_check_168():
-  """ The variable font 'wdth' (Width) axis coordinate
-      must be 100 on the 'Regular' instance. """
-  from fontbakery.specifications.fvar import com_google_fonts_check_168 as check
-  from fontbakery.specifications.shared_conditions import regular_wdth_coord
-
-  # Our reference varfont, CabinVFBeta.ttf, has
-  # a bad Regular:wdth coordinate
-  ttFont = TTFont("data/test/cabinvfbeta/CabinVFBeta.ttf")
-  regular_width_coord = regular_wdth_coord(ttFont)
-
-  # So it must FAIL the test
-  print('Test FAIL with a bad Regular:wdth coordinate...')
-  status, message = list(check(ttFont, regular_width_coord))[-1]
-  assert status == FAIL
-
-  # We then fix the value:
-  ttFont["fvar"].instances[0].coordinates["wdth"] = 100
-  # Note: I know the correct instance index for this hotfix because
-  # I inspected the our reference CabinVF using ttx
-
-  # Then re-read the coord:
-  regular_width_coord = regular_wdth_coord(ttFont)
-
-  # and now this should PASS the test:
-  print('Test PASS with a good Regular:wdth coordinate (100)...')
-  status, message = list(check(ttFont, regular_width_coord))[-1]
-  assert status == PASS
-
-
-
-def test_check_169():
-  """ The variable font 'slnt' (Slant) axis coordinate
-      must be zero on the 'Regular' instance. """
-  from fontbakery.specifications.fvar import com_google_fonts_check_169 as check
-  from fontbakery.specifications.shared_conditions import regular_slnt_coord
-  from fontTools.ttLib.tables._f_v_a_r import Axis
-
-  # Our reference varfont, CabinVFBeta.ttf, lacks a 'slnt' variation axis.
-  ttFont = TTFont("data/test/cabinvfbeta/CabinVFBeta.ttf")
-
-  # So we add one:
-  new_axis = Axis()
-  new_axis.axisTag = "slnt"
-  ttFont["fvar"].axes.append(new_axis)
-
-  # and specify a bad coordinate for the Regular:
-  ttFont["fvar"].instances[0].coordinates["slnt"] = 123
-  # Note: I know the correct instance index for this hotfix because
-  # I inspected the our reference CabinVF using ttx
-
-  # then we test the code of the regular_slnt_coord condition:
-  regular_slant_coord = regular_slnt_coord(ttFont)
-
-  # And with this the test must FAIL
-  print('Test FAIL with a bad Regular:slnt coordinate (123)...')
-  status, message = list(check(ttFont, regular_slant_coord))[-1]
-  assert status == FAIL
-
-  # We then fix the Regular:slnt coordinate value value:
-  regular_slant_coord = 0
-
-  # and now this should PASS the test:
-  print('Test PASS with a good Regular:slnt coordinate (zero)...')
-  status, message = list(check(ttFont, regular_slant_coord))[-1]
-  assert status == PASS
-
-
-def test_check_170():
-  """ The variable font 'ital' (Italic) axis coordinate
-      must be zero on the 'Regular' instance. """
-  from fontbakery.specifications.fvar import com_google_fonts_check_170 as check
-  from fontbakery.specifications.shared_conditions import regular_ital_coord
-  from fontTools.ttLib.tables._f_v_a_r import Axis
-
-  # Our reference varfont, CabinVFBeta.ttf, lacks an 'ital' variation axis.
-  ttFont = TTFont("data/test/cabinvfbeta/CabinVFBeta.ttf")
-
-  # So we add one:
-  new_axis = Axis()
-  new_axis.axisTag = "ital"
-  ttFont["fvar"].axes.append(new_axis)
-
-  # and specify a bad coordinate for the Regular:
-  ttFont["fvar"].instances[0].coordinates["ital"] = 123
-  # Note: I know the correct instance index for this hotfix because
-  # I inspected the our reference CabinVF using ttx
-
-  # then we test the code of the regular_ital_coord condition:
-  regular_italic_coord = regular_ital_coord(ttFont)
-
-  # So it must FAIL the test
-  print('Test FAIL with a bad Regular:ital coordinate (123)...')
-  status, message = list(check(ttFont, regular_italic_coord))[-1]
-  assert status == FAIL
-
-  # We then fix the Regular:ital coordinate:
-  regular_italic_coord = 0
-
-  # and now this should PASS the test:
-  print('Test PASS with a good Regular:ital coordinate (zero)...')
-  status, message = list(check(ttFont, regular_italic_coord))[-1]
-  assert status == PASS
-
-
-def test_check_171():
-  """ The variable font 'opsz' (Optical Size) axis coordinate
-      should be between 9 and 13 on the 'Regular' instance. """
-  from fontbakery.specifications.fvar import com_google_fonts_check_171 as check
-  from fontbakery.specifications.shared_conditions import regular_opsz_coord
-  from fontTools.ttLib.tables._f_v_a_r import Axis
-
-  # Our reference varfont, CabinVFBeta.ttf, lacks an 'opsz' variation axis.
-  ttFont = TTFont("data/test/cabinvfbeta/CabinVFBeta.ttf")
-
-  # So we add one:
-  new_axis = Axis()
-  new_axis.axisTag = "opsz"
-  ttFont["fvar"].axes.append(new_axis)
-
-  # and specify a bad coordinate for the Regular:
-  ttFont["fvar"].instances[0].coordinates["opsz"] = 8
-  # Note: I know the correct instance index for this hotfix because
-  # I inspected the our reference CabinVF using ttx
-
-  # then we test the regular_opsz_coord condition:
-  regular_opticalsize_coord = regular_opsz_coord(ttFont)
-
-  # And it must WARN the test
-  print('Test WARN with a bad Regular:opsz coordinate (8)...')
-  status, message = list(check(ttFont, regular_opticalsize_coord))[-1]
-  assert status == WARN
-
-  # We try yet another bad value
-  regualr_opticalsize_coord = 14
-
-  # And it must also WARN the test
-  print('Test WARN with another bad Regular:opsz value (14)...')
-  status, message = list(check(ttFont, regular_opticalsize_coord))[-1]
-  assert status == WARN
-
-  # We then test with good default opsz values:
-  for value in [9, 10, 11, 12, 13]:
-    regular_opticalsize_coord = value
-
-    # and now this should PASS the test:
-    print('Test PASS with a good Regular:opsz coordinate ({})...'.format(value))
-    status, message = list(check(ttFont, regular_opticalsize_coord))[-1]
-    assert status == PASS
-
-
-def test_check_172():
-  """ The variable font 'wght' (Weight) axis coordinate
-      must be 700 on the 'Bold' instance. """
-  from fontbakery.specifications.fvar import com_google_fonts_check_172 as check
-  from fontbakery.specifications.shared_conditions import bold_wght_coord
-
-  # Our reference varfont, CabinVFBeta.ttf, has
-  # a bad Bold:wght coordinate
-  ttFont = TTFont("data/test/cabinvfbeta/CabinVFBeta.ttf")
-  bold_weight_coord = bold_wght_coord(ttFont)
-
-  # So it must FAIL the test
-  print('Test FAIL with a bad Bold:wght coordinate...')
-  status, message = list(check(ttFont, bold_weight_coord))[-1]
-  assert status == FAIL
-
-  # We then fix the value:
-  ttFont["fvar"].instances[3].coordinates["wght"] = 700
-  # Note: I know the correct instance index for this hotfix because
-  # I inspected the our reference CabinVF using ttx
-
-  # Then re-read the coord:
-  bold_weight_coord = bold_wght_coord(ttFont)
-
-  # and now this should PASS the test:
-  print('Test PASS with a good Bold:wght coordinage (700)...')
-  status, message = list(check(ttFont, bold_weight_coord))[-1]
-  assert status == PASS
 
 # Temporarily disabling this code-test since check/173 itself
 # is disabled waiting for an implementation targetting the
