+++ Lib/fontbakery/specifications/head_test.py
@@ -1,0 +1,75 @@
+# -*- coding: utf-8 -*-
+from __future__ import absolute_import, print_function, unicode_literals, division
+
+import pytest
+from fontbakery.checkrunner import (
+              DEBUG
+            , INFO
+            , WARN
+            , ERROR
+            , SKIP
+            , PASS
+            , FAIL
+            )
+
+check_statuses = (ERROR, FAIL, SKIP, PASS, WARN, INFO, DEBUG)
+
+from fontTools.ttLib import TTFont
+
+mada_fonts = [
+  "data/test/mada/Mada-Black.ttf",
+  "data/test/mada/Mada-ExtraLight.ttf",
+  "data/test/mada/Mada-Medium.ttf",
+  "data/test/mada/Mada-SemiBold.ttf",
+  "data/test/mada/Mada-Bold.ttf",
+  "data/test/mada/Mada-Light.ttf",
+  "data/test/mada/Mada-Regular.ttf",
+]
+
+@pytest.fixture
+def mada_ttFonts():
+  return [TTFont(path) for path in mada_fonts]
+
+
+def test_check_014(mada_ttFonts):
+  """ Make sure all font files have the same version value. """
+  from fontbakery.specifications.head import com_google_fonts_check_014 as check
+
+  print('Test PASS with good family.')
+  # our reference Mada family is know to be good here.
+  status, message = list(check(mada_ttFonts))[-1]
+  assert status == PASS
+
+  bad_ttFonts = mada_ttFonts
+  # introduce a mismatching version value into the second font file:
+  version = bad_ttFonts[0]['head'].fontRevision
+  bad_ttFonts[1]['head'].fontRevision = version + 1
+
+  print('Test WARN with fonts that diverge on the fontRevision field value.')
+  status, message = list(check(bad_ttFonts))[-1]
+  assert status == WARN
+
+def test_check_043():
+  """ Checking unitsPerEm value is reasonable. """
+  from fontbakery.specifications.head import com_google_fonts_check_043 as check
+
+  # In this test we'll forge several known-good and known-bad values.
+  # We'll use Mada Regular to start with:
+  ttFont = TTFont("data/test/mada/Mada-Regular.ttf")
+
+  for good_value in [1000, 16, 32, 64, 128, 256,
+                     512, 1024, 2048, 4096, 8192, 16384]:
+    print("Test PASS with a good value of unitsPerEm = {} ...".format(good_value))
+    ttFont['head'].unitsPerEm = good_value
+    status, message = list(check(ttFont))[-1]
+    assert status == PASS
+
+  # These are arbitrarily chosen bad values:
+  for bad_value in [0, 1, 2, 4, 8, 10, 100, 10000, 32768]:
+    print("Test FAIL with a bad value of unitsPerEm = {} ...".format(bad_value))
+    ttFont['head'].unitsPerEm = bad_value
+    status, message = list(check(ttFont))[-1]
+    assert status == FAIL
+
+# TODO: test_check_044
+
